refactor!: update import paths for logging and asset collection modules
BREAKING CHANGE:
- `hyperion.logging` renamed to `hyperion.log`
- `hyperion.infrastructure.queue` renamed to `hyperion.infrastructure.message_queue`
- `hyperion.infrastructure.secrets` renamed to `hyperion.infrastructure.secretsmanager`
- `hyperion.infrastructure.http` renamed to `hyperion.infrastructure.httputils`
- `hyperion.collections` moved to `hyperion.repository`
