+++ hyperion/asyncutils.py
@@ -4,7 +4,7 @@ import sys
 from collections.abc import AsyncIterable, AsyncIterator, Callable, Coroutine, Iterable
 from typing import TYPE_CHECKING, Any, Generic, TypeVar
 
-from hyperion.logging import get_logger
+from hyperion.log import get_logger
 
 if sys.version_info >= (3, 11) and TYPE_CHECKING:
     from typing import Self
