+++ hyperion/infrastructure/geo/location.py
@@ -12,7 +12,7 @@ import numpy.typing as npt
 from haversine.haversine import get_avg_earth_radius
 
 from hyperion.infrastructure.cache import Cache
-from hyperion.logging import get_logger
+from hyperion.log import get_logger
 
 LATITUDE_DEGREE_TO_METERS = 111_000
 EARTH_RADIUS_METERS = get_avg_earth_radius(haversine.Unit.METERS)
